pyproject.toml
[tool.poetry]
name = "crpt-test-assignment"
version = "0.1.0"
description = ""
authors = ["Vyacheslav Ratushnyi <ratushnyyvm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
